requirements.txt
ccxt
pyyaml
colorama
